2 versions of neuron.c with different time_step_update stuff
